requirements.txt
psutil
python-dotenv
requests
beautifulsoup4
dnspython
selenium
undetected-chromedriver
webdriver-manager
groq
openai
yelpapi
overpy
pydantic
langgraph
langchain-groq
httpx
fastapi
websockets
capsolver
google-search-results
proxycurl-py[asyncio]
